solution/src/main/java/backendcodingchallenge/service/serializers/JsonDateDeserializer.java
package backendcodingchallenge.service.serializers;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;
import org.springframework.stereotype.Component;

import java.io.IOException;
import java.text.ParseException;
import java.text.SimpleDateFormat;
import java.util.Date;

@Component
public class JsonDateDeserializer extends JsonDeserializer<Date> {

    @Override
    public Date deserialize(JsonParser jsonparser,
                            DeserializationContext deserializationcontext) throws IOException {


        SimpleDateFormat format = new SimpleDateFormat(JsonDateSerializer.FORMAT);
        format.setLenient(false); //Let's not accept "310/24/2010" and convert it in a somewhat-related date
        String date = jsonparser.getText();

        try {
            return format.parse(date);
        } catch (ParseException e) {
            throw new RuntimeException(e);
        }

    }
}